refactor: OAuthConfig と createOAuthConfig を shared から adapter 層に移動
OAuth 設定型と設定生成関数が shared レイヤーにあり、アダプタ層の内部詳細が
共有レイヤーに露出していた。adapter/chrome/ に移動してカプセル化を強化。
