+++ src/adapter/chrome/identity.adapter.ts
@@ -1,8 +1,8 @@
 import type { AuthPort } from "../../domain/ports/auth.port";
 import type { StoragePort } from "../../domain/ports/storage.port";
 import type { AuthToken, DeviceCodeResponse, PollResult } from "../../domain/types/auth";
-import type { OAuthConfig } from "../../shared/types/auth";
 import { AuthError, isAuthToken } from "../../shared/types/auth";
+import type { OAuthConfig } from "./oauth.config";
 
 const TOKEN_STORAGE_KEY = "github_auth_token";
